+++ controllers/StoryController.php
@@ -9,6 +9,7 @@ namespace app\controllers;
 
 //Imports
 use app\components\SafeToolController;
+use app\models\ProductOwner;
 use app\models\Story;
 use app\models\StoryAcceptanceCriteria;
 use Exception;
@@ -47,7 +48,7 @@ class StoryController extends SafeToolController
 		$model = $this->findModel($id);
 		$searchModel = new StoryAcceptanceCriteria();
 		$dataProvider = $searchModel->search(Yii::$app->getRequest()->getQueryParams(), $model->getAttribute('id'));
-		
+
 		return $this->render('view', [
 			'model' => $model,
 			'dataProvider' => $dataProvider,
@@ -60,10 +61,12 @@ class StoryController extends SafeToolController
 	 * If creation is successful, the browser will be redirected to the 'view' page.
 	 *
 	 * @return string
+	 * @throws NotFoundHttpException
 	 */
 	public function actionCreate()
 	{
 		$model = new Story();
+		$model->setAttribute('product_owner_id', ProductOwner::getCurrentProductOwnerId());
 
 		if ($model->load(Yii::$app->getRequest()->post()) && $model->save()) {
 			return $this->redirect(['view', 'id' => $model->getAttribute('id')]);
@@ -86,10 +89,42 @@ class StoryController extends SafeToolController
 	{
 		$model = $this->findModel($id);
 
+		if (ProductOwner::getCurrentProductOwnerId() != $model->getAttribute('product_owner_id')) {
+			throw new NotFoundHttpException(Yii::t('story',
+				'Only the product owner responsible for this story can change it.'));
+		}
+
 		if ($model->load(Yii::$app->getRequest()->post()) && $model->save()) {
 			return $this->redirect(['view', 'id' => $model->getAttribute('id')]);
 		} else {
 			return $this->render('update', [
+				'model' => $model,
+			]);
+		}
+	}
+
+	/**
+	 * Transfer the product owner of the story.
+	 * If update is successful, the browser will be redirected to the 'view' page.
+	 *
+	 * @param integer $id Story ID.
+	 * @return string
+	 * @throws NotFoundHttpException
+	 */
+	public function actionTransfer($id)
+	{
+		$model = $this->findModel($id);
+		$model->setScenario('transfer');
+
+		if (ProductOwner::getCurrentProductOwnerId() != $model->getAttribute('product_owner_id')) {
+			throw new NotFoundHttpException(Yii::t('story',
+				'Only the product owner responsible for this story can transfer it.'));
+		}
+
+		if ($model->load(Yii::$app->getRequest()->post()) && $model->save()) {
+			return $this->redirect(['view', 'id' => $model->getAttribute('id')]);
+		} else {
+			return $this->render('transfer', [
 				'model' => $model,
 			]);
 		}
@@ -108,6 +143,12 @@ class StoryController extends SafeToolController
 	public function actionDelete($id)
 	{
 		$model = $this->findModel($id);
+
+		if (ProductOwner::getCurrentProductOwnerId() != $model->getAttribute('product_owner_id')) {
+			throw new NotFoundHttpException(Yii::t('story',
+				'Only the product owner responsible for this story can delete it.'));
+		}
+
 		try {
 			$model->delete();
 		} catch (Exception $ex) {
